FastAPI_ML/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
